requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cffi==1.15.1
charset-normalizer==2.1.1
clr-loader==0.1.7
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
proxy-tools==0.1.0
pycparser==2.21
pythonnet==3.0.0a2
pywebview==3.6.3
yarl==1.8.1
